+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "dbapi",
+  "version": "0.0.1",
+  "description": "An API to connect to our database",
+  "main": "./build/main.js",
+  "scripts": {
+    "start": "tsc ./src/main.ts --outDir ./build && node ./build/main.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Coders-Bubble/dbAPI.git"
+  },
+  "author": "",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Coders-Bubble/dbAPI/issues"
+  },
+  "homepage": "https://github.com/Coders-Bubble/dbAPI#readme",
+  "dependencies": {
+    "express": "latest",
+    "typescript": "latest",
+    "mysql": "latest",
+    "@types/express": "latest",
+    "@types/node": "latest",
+    "@types/mysql": "latest"
+  },
+  "devDependencies": {
+    "eslint": "latest"
+  }
+}
